refactor: abstract saveFile and formatTypeScript utils
