+++ src/utils/index.ts
@@ -1,5 +1,4 @@
 export { patchJsonSchema } from './patchJsonSchema';
-export { clearFolder } from './clearFolder';
 export { jsonSchemaToTsConst } from './jsonSchemaToTsConst';
 export { convertOpenApiParameters } from './convertOpenApiParameters';
 export { convertOpenApiToJsonSchema } from './convertOpenApiToJsonSchema';
@@ -14,4 +13,8 @@ export {
 export { replaceInlinedRefsWithStringPlaceholder } from './replaceInlinedRefsWithStringPlaceholder';
 export { replacePlaceholdersWithImportedSchemas } from './jsonSchemaToTsConst/replacePlaceholdersWithImportedSchemas';
 export { addSchemaToMetaData } from './addSchemaToMetaData';
+
+export { clearFolder } from './clearFolder';
 export { makeRelativePath } from './makeRelativePath';
+export { formatTypeScript } from './formatTypeScript';
+export { saveFile } from './saveFile';
