+++ src/utils/makeRelativePath.ts
@@ -1,4 +1,4 @@
-import path from 'path';
+import path from 'node:path';
 
 /**
  * Evaluate the relative path from/to the given absolute paths
