+++ test/test-utils/index.ts
@@ -1,2 +1,1 @@
 export { importFresh } from './importFresh';
-export { formatTypeScript } from './formatTypeScript';
